requirements.txt
psycopg2==2.8.5
django==3.0.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-filter==2.3.0
django-cors-headers==3.4.0
django-allauth==0.42.0
django-rest-auth==0.9.5